Curso Finalizado, se logra crear satisfactoriamente el arbol DOM, falta persistencia de datos
