+++ 1821-manipulando-dom-con-js-proyecto_inicial/components/deleteIcon.js
@@ -1,0 +1,15 @@
+const deleteIcon = () => {
+    const i = document.createElement('i')
+    i.classList.add("fas", "fa-trash-alt", "trashIcon", "icon");
+    i.addEventListener("click", deleteTask)
+    return i;
+};
+
+const deleteTask = (event) => {
+    //console.log("Eliminar tarea");
+    //console.log(event.target.parentElement);
+    const parent = event.target.parentElement;
+    parent.remove();
+};
+
+export default deleteIcon
