+++ 1821-manipulando-dom-con-js-proyecto_inicial/index.html
@@ -41,6 +41,6 @@
         </ul>
       </div>
     </main>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
   </body>
 </html>
