+++ 1821-manipulando-dom-con-js-proyecto_inicial/script.js
@@ -1,5 +1,7 @@
-( () => {
-    const btn = document.querySelector("[data-form-btn]");
+import checkComplete from "./components/checkComplete.js";
+import deleteIcon from "./components/deleteIcon.js";
+
+const btn = document.querySelector("[data-form-btn]");
 
 const createTask = (evento) => {
     evento.preventDefault();
@@ -15,31 +17,13 @@ const createTask = (evento) => {
     titleTask.innerText = value;
     taskContent.appendChild(checkComplete());
     taskContent.appendChild(titleTask);
-
-    const content = `
-    <i class="fas fa-trash-alt trashIcon icon"></i>`;
-
     //task.innerHTML = content;
     task.appendChild(taskContent);
+    task.appendChild(deleteIcon())
 
     list.appendChild(task)
-    console.log(content);
+    //console.log(content);
 };
-console.log(btn);
+//console.log(btn);
 
 btn.addEventListener('click', createTask);
-
-const checkComplete = () => {
-    const i = document.createElement('i');
-    i.classList.add("far", "fa-check-square", "icon");
-    i.addEventListener('click', completeTask);
-    return i;
-}
-//Inmediately invoked function expression IIFE
-const completeTask = (event) => {
-    const element = event.target;
-    element.classList.toggle("fas");
-    element.classList.toggle("completeIcon");
-    element.classList.toggle("far");
-} 
-})();
